plumbing: transport, move common and server transports
Git file, git, and ssh transports use a full-duplex transport protocol.
This is now located under `plumbing/transport`.
The server transport along with its `Serve*` methods are now located
under `plumbing/server`.
