+++ plumbing/server/loader_test.go
