+++ plumbing/transport/git/receive_pack_test.go
@@ -1,7 +1,7 @@
 package git
 
 import (
-	"github.com/go-git/go-git/v5/plumbing/transport/test"
+	"github.com/go-git/go-git/v5/internal/transport/test"
 
 	fixtures "github.com/go-git/go-git-fixtures/v4"
 	. "gopkg.in/check.v1"
