+++ plumbing/transport/http/upload_pack.go
@@ -11,7 +11,6 @@ import (
 	"github.com/go-git/go-git/v5/plumbing/format/pktline"
 	"github.com/go-git/go-git/v5/plumbing/protocol/packp"
 	"github.com/go-git/go-git/v5/plumbing/transport"
-	"github.com/go-git/go-git/v5/plumbing/transport/internal/common"
 	"github.com/go-git/go-git/v5/utils/ioutil"
 )
 
@@ -69,7 +68,7 @@ func (s *upSession) UploadPack(
 	}
 
 	rc := ioutil.NewReadCloser(r, res.Body)
-	return common.DecodeUploadPackResponse(rc, req)
+	return transport.DecodeUploadPackResponse(rc, req)
 }
 
 // Close does nothing.
