+++ plumbing/transport/ssh/proxy_test.go
@@ -10,8 +10,8 @@ import (
 
 	"github.com/armon/go-socks5"
 	"github.com/gliderlabs/ssh"
+	"github.com/go-git/go-git/v5/internal/transport/ssh/test"
 	"github.com/go-git/go-git/v5/plumbing/transport"
-	"github.com/go-git/go-git/v5/plumbing/transport/ssh/internal/test"
 
 	fixtures "github.com/go-git/go-git-fixtures/v4"
 	stdssh "golang.org/x/crypto/ssh"
