+++ plumbing/transport/transport_fuzz_test.go
@@ -1,0 +1,11 @@
+package transport
+
+import (
+	"testing"
+)
+
+func FuzzNewEndpoint(f *testing.F) {
+	f.Fuzz(func(_ *testing.T, input string) {
+		NewEndpoint(input)
+	})
+}
